+++ lib/config.rb
@@ -53,8 +53,9 @@ module Config
 
   # Loads and sets the settings constant!
   def self.load_and_set_settings(*files)
-    Kernel.send(:remove_const, Config.const_name) if Kernel.const_defined?(Config.const_name)
-    Kernel.const_set(Config.const_name, Config.load_files(files))
+    name = Config.const_name
+    Object.send(:remove_const, name) if Object.const_defined?(name)
+    Object.const_set(name, Config.load_files(files))
   end
 
   def self.setting_files(config_root, env)
@@ -70,7 +71,7 @@ module Config
   end
 
   def self.reload!
-    Kernel.const_get(Config.const_name).reload!
+    Object.const_get(Config.const_name).reload!
   end
 end